Fixed camera view in BucketView. Added posibility to add a drop to your profile. Refactored Cameraview, Started on the Info functionality in Bucket View

+++ wave/app/ActivityViewController.m
@@ -87,7 +87,7 @@ const int EXPAND_SIZE = 400;
         
         [CATransaction setCompletionBlock:^{
             if(shouldExpand){
-                [self expandBucketWithId:indexPath.row];
+                [self expandBucketWithId:(int)indexPath.row];
             }
             
         }];
